Combine SMOTE samplers into one Oversampler and move selection logic to fit()
